bookmark and office parsing support
