+++ artemis-docs/docs/Artifacts/overview.md
@@ -19,6 +19,6 @@ A breakdown of artifacts by OS is below.
 | OS                               | Number of Artifacts |
 | -------------------------------- | ------------------- |
 | [Windows](./windows.md)          | 33                  |
-| [macOS](./macos.md)              | 36                  |
+| [macOS](./macos.md)              | 37                  |
 | [Linux](./linux.md)              | 13                  |
-| [Application](./applications.md) | 6                   |
+| [Application](./applications.md) | 7                   |
